requirements.txt
Django
Pillow
django-cleanup 
django-unfold
python-dotenv